spark_app/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
pyspark>=4.0.0